+++ src/mnemonic.cpp
@@ -3,8 +3,7 @@
 #include "mnemonic_bit.hpp"
 #include "mnemonic_branch.hpp"
 #include "mnemonic_calc.hpp"
-#include "mnemonic_defb.hpp"
-#include "mnemonic_defw.hpp"
+#include "mnemonic_data.hpp"
 #include "mnemonic_ex.hpp"
 #include "mnemonic_increment.hpp"
 #include "mnemonic_io.hpp"
